requirements.txt
ccxt
pandas
google-generativeai
python-dotenv
feedparser
streamlit
requests
